Centered the linkedin icon
I set the Linkedin icon at the center of the rectangle. Also I little bit lowered the location of images of alumnus because compared to the leadership page, we don't have their position title and other links and the card looked too empty. By lowering the images, the card looks full.